// ui/src/renderer/src/houston/generated/McpToolState.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { AgentKind } from "./AgentKind";
import type { McpServer } from "./McpServer";

export type McpToolState = { tool: AgentKind, path: string, detected: boolean, servers: Array<McpServer>, error: string | null, };